scrapping/requirements.txt
# Core Scraping Framework
scrapy>=2.11.0
scrapy-user-agents>=0.1.1

# HTTP Clients
httpx>=0.26.0
requests>=2.31.0

# Data Processing
pandas>=2.1.0
pyarrow>=14.0.0
geopandas>=0.14.0
shapely>=2.0.0
networkx>=3.2

# PDF Processing
pdfplumber>=0.10.3
pytesseract>=0.3.10
pdf2image>=1.16.3
Pillow>=10.1.0

# Data Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Utilities
tenacity>=8.2.3
aiofiles>=23.2.1
python-dotenv>=1.0.0
tqdm>=4.66.0

# Logging
structlog>=23.2.0
python-json-logger>=2.0.7

# MCP Server
fastmcp>=0.2.0
mcp>=0.9.0

# Geospatial
fiona>=1.9.5
pyproj>=3.6.0
rtree>=1.1.0

# Web Parsing
beautifulsoup4>=4.12.2
lxml>=4.9.3

# Database for State Management
sqlalchemy>=2.0.23

# Rate Limiting
ratelimit>=2.2.1

# Testing (optional)
pytest>=7.4.3
pytest-asyncio>=0.21.1
